refactor: 기물은 기본적으로 한 수 단위로 움직인다
