+++ src/main/java/chess/domain/piece/King.java
@@ -6,7 +6,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-public class King extends FinitePiece {
+public class King extends QuadrantPiece {
 
     private King(boolean isWhite, Set<Move> moves) {
         super(isWhite, moves);
