+++ src/test/java/chess/domain/piece/PieceTest.java
@@ -1,5 +1,8 @@
 package chess.domain.piece;
 
+import static chess.domain.Direction.LEFT;
+import static chess.domain.Direction.RIGHT;
+import static chess.domain.Direction.UP;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.AbstractTestFixture;
@@ -15,11 +18,6 @@ public class PieceTest extends AbstractTestFixture {
         public PieceImplement(boolean isWhite, Set<Move> moves) {
             super(isWhite, moves);
         }
-
-        @Override
-        protected boolean compareMove(Move pieceMove, Move move) {
-            return false;
-        }
     }
 
     @DisplayName("색을 구별할 수 있다")
@@ -30,6 +28,35 @@ public class PieceTest extends AbstractTestFixture {
 
         assertThat(piece.hasSameColor(otherPiece)).isFalse();
     }
+
+    @DisplayName("가능한 수인지 판단한다(유한)")
+    @Test
+    void isValidMoveFinite() {
+        Move move = createMove(UP, UP, RIGHT);
+        Move move2 = createMove(RIGHT, RIGHT, UP);
+
+        Piece piece = createPiece(true, move, move2);
+        assertThat(piece.canMove(move)).isTrue();
+    }
+
+    @DisplayName("가능하지 않은 수인지 판단한다(기물:유한, 수:무한)")
+    @Test
+    void isInvalidMoveFinite() {
+        Move move = createMove(UP, RIGHT);
+
+        Piece piece = createPiece(true, move);
+        assertThat(piece.canMove(createMove(UP, RIGHT, UP, RIGHT))).isFalse();
+    }
+
+    @DisplayName("가능하지 않은 수인지 판단한다")
+    @Test
+    void isInvalidMove() {
+        Move move = createMove(UP, RIGHT);
+
+        Piece piece = createPiece(true, move);
+        assertThat(piece.canMove(createMove(LEFT))).isFalse();
+    }
+
 
     private Piece createPiece(boolean isWhite, Move... moves) {
         return new PieceTest.PieceImplement(isWhite, Set.of(moves));
